Renaming .c and .h files in order to make the whole thing compile correctly
