+++ keyboards/manyboard/macro/macro.c
@@ -14,5 +14,4 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-#include "manyboard_macro.h"
-
+#include "macro.h"
